SERVER-4008 -- don't leak memory in the shell
Track the strdup-ed (i.e. malloc-ed) strings returned by linenoise, and
free the memory when we are done with it. Free command line history
memory explicitly when we save the history to a file and stop using it.
Non-Windows versions were freeing this memory using an atexit handler,
but this handler isn't set up for Windows.


Former-commit-id: 8d8b272225b9a2e98f25525e0e2e7f032420388f