SERVER-4011 don't look to Esc to revert an active command completion
The code to accept Esc to revert an in-progress command completion
only worked in Windows, and it won't work on any platform when I
add code to accept Meta commands that can begin with Esc (on any
platform).  So, remove the code that looked for Esc during command
completion processing.


Former-commit-id: 35c76c0a277f3657685e064adae401e0e16607c3