shut down mongobridges explicitly, at the end of a replsettest
This is for Windows where we do not receive a signal when the socket
is shut down (and we're not blocked on the socket so we never notice
it's closed).
This only comes into play when you have multiple bridges to one mongod;
the bridge used to shut down the mongod will exit, but none of the
others will.


Former-commit-id: c5b2accea82c20d16f0e9fe8fd96b0db80b726da