Roll back most of the following two patches:
80da393875b95da15e5193b95887bf6540ba31e5 [formerly c4851043c682c9e2a9230166045ce85b518b85eb]:  qlock can ask if in greedy mode or not
6618531fdb6142ed72373ea6f26fac1a856215ab [formerly 938ac1619dc88dc5f9e6df47dcb981fbf0f33663]: if in ThreadSpanningOp don't profile

The problem is that isGreedSuspended cannot be called with the qlock's internal
mutex already held.  However, a caller not holding that lock already can't
know that by the time the function returns, its return value is still valid.
Another thread may change it in parallel.  The result is that the attempt
to use it to stop profiling when thread-spanning ops are enabled only appears
to provide that protection.


Former-commit-id: b8cec6e3c0df43c63e096dabd93d5697c6143b82