SERVER-6228 treat inserts into .system.indexes as commands
Index builds are not replicated as a 'c' op (command) but instead
as an 'i' op (insert) into the .system.indexes collection. The
insert code detects this and builds the index as part of the insert
of the record.
To make multithreaded replication work, we need to treat index builds
like commands so that they aren't intermixed with other write ops in
a batch.


Former-commit-id: 028db4bb3ec21677033924100cb2dcd96678dadf