Fix use of ByLocKey as an incomplete type in standard containers.
ByLocKey needs to be a complete type in database.h since it is used in
a std::map, and it is illegal in C++03 to use an incomplete type in a
standard container. We can't just include clientcursor.h in database.h
though, as that would be circular, so move ByLocKey to a new header.


Former-commit-id: e2a1f7522706ee0eb7e83b48b2fa85fda31f54c4