SERVER-8017 Improve logic and logging in awaitReplication()
Do not print "ReplSetTest [object Object]" at the start of a run.  Do not continue
to check other secondaries in the callback once one has been found to be stale.  Do
not silently reset the timestamp we are checking against and continue checking with
a different timestanp.  Do not print unlabeled oplog entries.  Do not print diagnostic
information when a secondary is found to be in sync.  Display success or failure
indication for each checked secondary, and include diagnostic information on failure.
Display updated timestamp and primary name when recovering after an exception.  Display
success message on success.  Label logged lines consistently so they can be searched
for in logs.


Former-commit-id: a321943143ef35ae26893b98847207c2045703e6