SERVER-8414/SERVER-8429 Support SASL authentication in DBClient::auth() method, tools.
Introduces a new overload of the auth() method of DBClientWithCommands to the
C++ driver.  The new overload takes a single BSONObj of authentication
parameters as input, and is analagous to the saslClientAuthenticate method.
This new method supports all SASL methods, if the client has GSASL support and
previously ran the mongo initializers, plus it supports MONGO_CR.  It is the
preferred entry point for authentication in the C++ client, going forward.

The old ::auth() method remains and keeps its existing behavior.


Former-commit-id: 1b2b144a8829296cfb1a8f78864eebfe752d9cb3