Visual Studio -- add JavaScript source code to all projects
Since the JavaScript source files are included in all projects,
it is inconvenient to have them listed in only the mongo (shell)
project, so add them to all other projects.  Split the list into
a common (all projects) group and a shell-only group.  Mark the
unused (in a particular project) ones as not-included.  Add
missing assert.js and types.js files to the lists.


Former-commit-id: 8365d3949be3e877fb73ebb25a8c1440b8bb3b40