SCons: Make glibc check a function of --release flag, not presence of --s3dist on command line.
Release builds of mongo on Linux are not supposed to depend on symbols from
GLIBC2.4.  Prior to this patch, we enforced this by adding a post build option
to Program steps on linux when s3dist appeared on the command line, unless
suppressed by --no-glibc-check.  In this patch, we instead add the post build
option if --release is set on the command line, unless suppressed by the
--no-glibc-check.  Without this change, the s3dist rule can cause unnecessary
relinking of the programs to be archived and pushed.


Former-commit-id: 0c1895dae4a9cd6550f8ce0a8c40fe4d988361d5