SERVER-8741 Add more diagnostics to test
Keep track of the shard key value (random number) for each saved
document and report it when a document can't be found so we can
see what shard it should be on and when it has been migrated.
When mongos can't find a document, look for it directly on both
shards and report where it was found or if it was not found.
Add header comments describing these features.


Former-commit-id: 3f2de68c7ee94db972ee4a631208c3c8b0ee6f8f