SERVER-9307 initialize session id context to support session caching on the server
You must set the context in order for session caching to work.
This would normally fail silently without setting the context, but when you turn on
certificate validation and attempt to reuse a session, the attempt will actively fail and
return an error on the server.
The Java driver does make use of restartable sessions and is affected by this issue.


Former-commit-id: 74409f3cb7a2122610842955ed1c6ed7dd1779e6