SERVER-1105 Use ResourcePattern type when identifying the resource component of required privileges.
This patch has two principal components.  First, it changes the interface to Privilege and
AuthorizationSession to use ResourcePattern in place of std::string for identifying resources.
Second, it examines all call sites of the authorization session interface in commands and
other code to ensure that the correct resource requirements are conveyed to the authorization_session.


Former-commit-id: a7f7c028c4cdda1ab0939c6c7788bb39bd94cc5f