SERVER-12896 MongoRunner fixes to make multiversion tests run on Windows
The system programs either come versioned in the form of
<utility>-<major.minor> (e.g., mongorestore-2.4) or just <utility>. If
change_extension is called on this string, it will replace the minor
version component and the binary will not be found.

In pre-2.6.0rc0 versions, we did not have the server shutdown event and
the stop commands needs to use the old method of connecting on the admin
port and invoking shutdown.


Former-commit-id: 32b36216dd85697c614c4873021387658d31e1a0