SERVER-11770 Support changing the default gssapiServiceName from the shell command line.
This patch allows the user to specify an alternate gssapiServiceName (default is mongodb)
when authenticating using GSSAPI/Kerberos.  It also allows the user to specify the host
name to use for authentication purposes, when this does not match the DNS host name.
Both of these functions were previously exposed only through the db.auth() method,
and not for command-line authentication.


Former-commit-id: 944807590b5d7a6a5b7b53f02bd032faf9406507