SERVER-13295 Ensure that user and data documents reach all nodes in auth_no_pri.js.
In order to confirm that you can authenticate to and read data from a
non-primary node with auth enabled, the test set up phase must write some data
and wait for it to reach all nodes, before taking down the majority of them.
This patch achieves this by setting an appropriate write concern on the user and
data document inserts.


Former-commit-id: 7001e916bd78097b34812c263274eaeb5fd712cf