SERVER-13635 Replaces high-level references to Record  with RecordData class.
In order to allow the use of certain types of storage engines, it was
necessary to replace some use of the Record class. The Record class is
now almost entirely restricted to the mmapV1 level and below.
Higher-level uses of the Record class were replaced with calls to the
new RecordData class.


Former-commit-id: 5bd73a2f5ae33168cb554bcb002c91433fafc047