SERVER-13961 Move clean shutdown logic from dbexit to exitCleanly only
All remaining places, which call dbexit, outside of repl are in response
to unhandled exceptions, so it is not safe to call commit.

The repl calls happen at secondary replica startup time, so no writes
would have happened, so there should not be any data to throw out outside
of replication bookkeeping data.


Former-commit-id: b9a211782b7ca536ccb6c23b2017434317ca3da3