SERVER-11290 Don't actually build static libraries
We don't currently use the static libraries, they are just nodes in our
dependency graph. Since they are unused, building them wastes time and disk
IO.

We still need to generate the files to appease scons. They are filled with a
UUID on the assumption that if scons thinks that a library needs to be rebuilt
then anything that depends on the library should also be rebuilt.


Former-commit-id: df27bbcf24db28fa68bfd264f0d3883c0e33320e