SERVER-15535 On reconfig, it should be OK to change the host associated with a given config id.
The new validation code erroneously forbade changing id -> host mapping.  It only needed to
forbid changing the host -> id mapping.  That is, it is all right to change the host representing
an id that appears in both configs, but it is not ok to change the id of a host that appears in
both configs.

Additionally, because we no longer hang up on non-step-down reconfigs, isself_different_port.js's
call to ReplSetTest.initiate is no longer expected to throw an exception.


Former-commit-id: 538bc36bfb6fde34d50e5c196c5c30d2e8ccaf71