SERVER-15781 Separate the concepts of next electable time and step-down until in replication.
The topology coordinator needs to track two distinct notions of a node not being
willing to stand for election.  The first is that of the step-down period.
During the step down period, the node will not stand for election due to the
step-down command being issued.  In heartbeat responses, nodes in the step down
period should report that they are not electable.

The second notion is that of the post election sleep period.  If a node loses an
election, it declines to stand again for a brief period of time (1 second plus a
random interval), to avert problems related to ties.  During this period, its
heartbeat responses should still report the node is electable, to suppress the
election of lower priority nodes.


Former-commit-id: 07c0c80e190206b8e6f037a885bc5971a78cae3e