SERVER-16500 Change the meaning of getReplicationMode() to be constant after start up.
Previously, ReplicationCoordinator::getReplicationMode() returned modeNone if
the settings object indicated the node was part of a replica set, but no valid
replica set configuration had yet been discovered.  With this change, the
settings object is the sole determinant of the return value of
getReplicationMode().  Callers that wish to know if a replica set member is
initialized should instead check ReplicationCoordinator::getMemberState().  If
the value is RS_STARTUP, then the node is not initialized with a replica set
config.

This has the side-effect of removing the mutex in ReplCoordinatorImpl from the
code path for this method.


Former-commit-id: 98a7e7f9709413ce06f49b48dee98f7080505947