SERVER-18131 Clean up LastError.
Makes LastError a decoration on Client.  Removes behavior that was specific to
the dbKillCursors wire protocol message into the processing for that message,
and out of last error.  Simplifies lifetime of LastError.  It always exists on
Clients, so there's no checking for its existence, and no passing it around
through parts of the networking library.


Former-commit-id: da31be34dc6664d76868543e9ccdaecad4f504c6