SERVER-14995 Move operation id, lockState and client fields to OperationContext.
They have been moved from OperationContextImpl.  Furthermore, the CurOp stack is
now attached to OperationContext, instead of Client.

With this change, an operation's lifetime is governed by the lifetime of an
OperationContext object.  The "_active" field of CurOp is therefore no longer
meaingful.  This required fixing the lifetime of OperationContext in a few
places.

A future change will adjust operation lifetime timing to time the lifetime of
the OperationContext object, as well.


Former-commit-id: 4ea38c308da292f43e29d32b1b53b7324db0bafe