SERVER-18917 make DBCommandCursor::next() shell method not require calling hasNext() first
In the current interface between the shell and the c++ driver, hasNext() is
what fetches the results from the server if at the end of the current batch.
next() just reads from the current batch. DBQuery smooths over this weirdness
by calling hasNext() from next() and throwing if it returns false.
DBCommandCursor should do the same.


Former-commit-id: 6acbfbba4f2e9e33221d71f2241642b6f9d2e479