SERVER-18464/SERVER-18437/SERVER-19004 Move targeter ownership to Shard
Currently there is a discrepancy between the ownership of the targeter
object and the lifetime of Shard. It is possible that if a shard has been
created and a targeter disappears then operations on that shard will not
find the targeter.

This change moves the ownership of targeter to the Shard object itself.


Former-commit-id: 6584bfd3ddcdb1598fb486615c49f0272798f5d6