SERVER-19855 Include min OpTime with shard version
This change adds the OpTime of the chunk manager as of the time its cached
metadata was loaded along with each versioned request sent from MongoS.
This includes write commands and the setShardVersion command.

The OpTime is only sent as part of this change. There will be a follow-up
change to add code to interpret this information on the MongoD side.


Former-commit-id: cec89b290906adb7d36507d20d619aaa5a6f6c16