SERVER-20234 Async MockStream's race in unblock
AsyncMockStreamFactor::MockStream::unblock() drops its lock to unblock
consuming threads. Then it re-acquires the lock before returning.

Under enough contention, this can lead simulateServer to attempt to
re-acquire a lock that has just gotten deleted if completeOperation
finishes within NetworkInterfaceASIO in time.

The solution is to avoid dropping the lock for the notify.


Former-commit-id: 78f7878663828b5fb70d102a338a32dd91ac2cea