SERVER-20764 Adjust create_index_gle.js to eliminate race between shell and mongos.
The problem with create_index_gle.js that causes SERVER-20764 was that
when the node ultimately given a slaveDelay is initially elected
primary, the test does not wait for mongos to recognize the change of
primary node in the shard replica set. The fix in this patch is to
initialize the shard replica set with only one electable node.


Former-commit-id: d9f2d8fda800fdc2cbd2be30413412c19a24f97a