SERVER-22867 Convey to access planner when to intersect bounds.
The plan enumerator is already aware of the rules for intersecting and
compounding bounds with multikey indexes. By making this information
part of the index assignment, the access planner can know whether it
should intersect the bounds or generate a self-intersection plan.

This change avoids creating duplicate work in the access planner that
already has to exist in the plan enumerator when the rules for
intersecting and compounding bounds becomes based on what prefixes of
indexed fields cause the index to be multikey.


Former-commit-id: 7467ccea16a70e514f5044866e56c8c4ebb0ad80