SERVER-22881 Expand arrays during $lookup's query.
If the value in 'localField' is an array, assume that it corresponds to
multiple entries in the foreign collection. The query to the foreign
collection will use a $in predicate, except when a regex is present, in
which case it will use a series of $eq predicates within an $or.


Former-commit-id: f23e588c180cc1a2ae79b9cbf6a4ba2a196ee215