SERVER-23609 Introduce RAII class for retrieving sharding cache objects
Around the sharding code there are a lot of repeated code sequences, where
we first check for existence of the database, then get the chunk manager.
In an effort to reduce this code duplication and to eventually be able to
get rid of the shared pointer references, this change introduces RAII
objects similar to the ones we have in mongod.


Former-commit-id: dff63f2afc63ee0b4e3288f50d435af83995e370